fix: resolve all pre-existing mypy type-check errors
- csm/matching.py: rename loop variables to avoid MatchCandidate/int
  type shadowing in HybridMechanism.match DA section
- csm/negotiation.py: cast arithmetic results to float for
  BoulwareStrategy and AdaptiveStrategy return values
- csm/runner.py: annotate matcher and strategy variables with base
  class types to allow polymorphic assignment
- metrics/rlm_metrics.py: annotate lags as list[float] to allow
  float("inf") appends

+++ swarm/csm/negotiation.py
@@ -69,11 +69,11 @@ class BoulwareStrategy(NegotiationStrategy):
         if is_buyer:
             lo = state.seller_reservation * 0.8
             hi = state.buyer_reservation
-            return lo + concession * (hi - lo)
+            return float(lo + concession * (hi - lo))
         else:
             hi = state.buyer_reservation * 1.2
             lo = state.seller_reservation
-            return hi - concession * (hi - lo)
+            return float(hi - concession * (hi - lo))
 
     def should_accept(
         self, state: NegotiationState, offer: float, is_buyer: bool
@@ -105,9 +105,9 @@ class AdaptiveStrategy(NegotiationStrategy):
             concession = 0.1 * t
 
         if is_buyer:
-            return state.buyer_reservation * (0.6 + concession)
+            return float(state.buyer_reservation * (0.6 + concession))
         else:
-            return state.seller_reservation * (1.4 - concession)
+            return float(state.seller_reservation * (1.4 - concession))
 
     def should_accept(
         self, state: NegotiationState, offer: float, is_buyer: bool
